requirements.txt
pandas
seaborn
matplotlib
opencv-python
fastapi
sqlalchemy
pytest
uvicorn
sqlalchemy-utils
python-dotenv
httpx
numpy
pandas
psycopg2-binary
jupyter
scikit-learn
passlib
argon2_cffi
PyJWT
joblib==1.3.2
imblearn
google-genai
pytest-mock